Ensured that all exceptions thrown during schema construction are caught and reported in the schemas panel (#164)
