+++ src/main/com/intellij/lang/jsgraphql/schema/GraphQLSchemaWithErrors.java
@@ -8,15 +8,11 @@
 package com.intellij.lang.jsgraphql.schema;
 
 import com.google.common.collect.Lists;
-import graphql.AssertException;
-import graphql.ErrorType;
 import graphql.GraphQLError;
 import graphql.GraphQLException;
-import graphql.language.SourceLocation;
 import graphql.schema.GraphQLSchema;
 import graphql.schema.idl.errors.SchemaProblem;
 
-import java.util.Collections;
 import java.util.List;
 
 public class GraphQLSchemaWithErrors {
@@ -56,24 +52,8 @@ public class GraphQLSchemaWithErrors {
                 errors.addAll(((SchemaProblem) exception).getErrors());
             } else if (exception instanceof GraphQLError) {
                 errors.add((GraphQLError) exception);
-            } else if (exception instanceof AssertException) {
-                errors.add(new GraphQLError() {
-                    @Override
-                    public String getMessage() {
-                        // strip out graphql-java internals part of the exception message given that the schema can actually be broken/invalid as the editor changes
-                        return exception.getMessage().replace("Internal error: should never happen: ", "");
-                    }
-
-                    @Override
-                    public List<SourceLocation> getLocations() {
-                        return Collections.emptyList();
-                    }
-
-                    @Override
-                    public ErrorType getErrorType() {
-                        return ErrorType.ValidationError;
-                    }
-                });
+            } else {
+                errors.add(new GraphQLInternalSchemaError(exception));
             }
         }
         return errors;
